requirements.txt
pandas
numpy
mlflow
notebook
scikit-learn
matplotlib
python-box
pyyaml
tqdm
ensure
joblib
types-pyyaml
Flask
Flask-Cors
ipykernel
